Robustify fMRIprep hack
To work araound a bug in fMRIprep, we move all json files in
a non-json-extension (else old version of fmriprep will attempt
to read them and fail if they aren't retrieved). Jean encountered
a BIDS-ish dataset without any jsons, where the find...mv
command failed. The --no-run-if-empty option of xargs safeguards
this, preventing mv to be run if the find command does not return
any results